gwsumm.channels: moved channel access to module
    - get_channel/get_channels moved to new module
