+++ bin/gw_summary
@@ -79,7 +79,7 @@ from gwpy.spectrum import psd
 
 from gwsumm import (globalv, version, mode, html)
 from gwsumm.config import *
-from gwsumm.data import get_channels
+from gwsumm.channels import get_channels
 from gwsumm.tabs import get_tab
 from gwsumm.utils import *
 from gwsumm.state import *
