+++ gwsumm/data.py
@@ -23,8 +23,6 @@ import operator
 import os
 import urllib2
 from math import (floor, ceil, pi, sqrt)
-from Queue import Queue
-import threading
 try:
     from configparser import (ConfigParser, NoSectionError, NoOptionError)
 except ImportError:
@@ -34,6 +32,7 @@ try:
 except ImportError:
     from ordereddict import OrderedDict
 
+import decorator
 import numpy
 import nds2
 import warnings
@@ -43,9 +42,10 @@ from astropy import units
 
 from glue import datafind
 from glue.lal import Cache
+from glue.segments import segmentlist
 
 from gwpy.detector import Channel
-from gwpy.segments import (DataQualityFlag, SegmentList)
+from gwpy.segments import (DataQualityFlag, SegmentList, Segment)
 from gwpy.timeseries import (TimeSeries, TimeSeriesList, TimeSeriesDict,
                              StateVector, StateVectorDict)
 from gwpy.spectrum import Spectrum
@@ -55,6 +55,7 @@ from gwpy.io import nds as ndsio
 from . import (globalv, version)
 from .mode import *
 from .utils import *
+from .channels import get_channel
 
 OPERATOR = {
     '*': operator.mul,
@@ -67,150 +68,18 @@ __author__ = 'Duncan Macleod <duncan.macleod@ligo.org>'
 __version__ = version.version
 
 
-class ThreadChannelQuery(threading.Thread):
-    """Threaded CIS `Channel` query.
+# -----------------------------------------------------------------------------
+# data access
+
+@decorator.decorator
+def use_segmentlist(f, arg1, segments, *args, **kwargs):
+    """Decorator a method to convert incoming segments into a `SegmentList`
     """
-    def __init__(self, inqueue, outqueue):
-        threading.Thread.__init__(self)
-        self.in_ = inqueue
-        self.out = outqueue
-
-    def run(self):
-        i, channel = self.in_.get()
-        self.in_.task_done()
-        try:
-            self.out.put((i, get_channel(channel, False)))
-        except Exception as e:
-            self.out.put(e)
-        self.out.task_done()
-
-
-def get_channel(channel, find_trend_source=True, timeout=5):
-    """Define a new :class:`~gwpy.detector.channel.Channel`
-
-    Parameters
-    ----------
-    channel : `str`
-        name of new channel
-
-    Returns
-    -------
-    Channel : :class:`~gwpy.detector.channel.Channel`
-        new channel.
-    """
-    if ' ' in str(channel):
-        name = str(channel)
-        try:
-            type_ = Channel.MATCH.match(name).groupdict()['type']
-        except AttributeError:
-            type_ = None
-        found = globalv.CHANNELS.sieve(name=name.replace('*', '\*'),
-                                       exact_match=True)
-    elif ',' in str(channel):
-        name, type_ = str(channel).rsplit(',', 1)
-        found = globalv.CHANNELS.sieve(name=name, type=type_, exact_match=True)
-    else:
-        type_ = isinstance(channel, Channel) and channel.type or None
-        sr = isinstance(channel, Channel) and channel.sample_rate or None
-        name = str(channel)
-        found = globalv.CHANNELS.sieve(name=str(channel), type=type_,
-                                       sample_rate=sr, exact_match=True)
-    if len(found) == 1:
-        return found[0]
-    elif len(found) > 1:
-        cstrings = ['%s [%s, %s]' % (c.ndsname, c.sample_rate, c.unit)
-                    for c in found]
-        raise ValueError("Ambiguous channel request '%s', multiple existing "
-                         "channels recovered:\n    %s"
-                         % (str(channel), '\n    '.join(cstrings)))
-    else:
-        matches = list(Channel.MATCH.finditer(name))
-        # match single raw channel
-        if len(matches) == 1 and not re.search('\.[a-z]+\Z', name):
-            try:
-                new = Channel.query(name, timeout=timeout)
-            except (ValueError, urllib2.URLError):
-                new = Channel(str(channel))
-            else:
-                new.name = str(channel)
-        # match single trend
-        elif len(matches) == 1:
-            # set default trend type based on mode
-            if type_ is None and globalv.MODE == SUMMARY_MODE_GPS:
-                type_ = 's-trend'
-            elif type_ is None:
-                type_ = 'm-trend'
-            name += ',%s' % type_
-            new = Channel(name)
-            if find_trend_source:
-                try:
-                    source = get_channel(new.name.rsplit('.')[0])
-                except ValueError:
-                    pass
-                else:
-                    new.url = source.url
-                    new.unit = source.unit
-                    try:
-                        new.bits = source.bits
-                    except AttributeError:
-                        pass
-                    try:
-                        new.filter = source.filter
-                    except AttributeError:
-                        pass
-            # determine sample rate for trends
-            if type_ == 'm-trend':
-                new.sample_rate = 1/60.
-            elif type_ == 's-trend':
-                new.sample_rate = 1
-        # match composite channel
-        else:
-            parts = get_channels([m.group() for m in matches])
-            new = Channel(name)
-            new.subchannels = parts
-            new._ifo = "".join(set(p.ifo for p in parts))
-        globalv.CHANNELS.append(new)
-        try:
-            return get_channel(new)
-        except RuntimeError as e:
-            if 'maximum recursion depth' in str(e):
-                raise RuntimeError("Recursion error while access channel "
-                                   "information for %s" % str(channel))
-            else:
-                raise
-
-
-def get_channels(channels):
-    """Multi-threaded channel query
-    """
-    if len(channels) == 0:
-        return []
-
-    # set up Queues
-    inqueue = Queue()
-    outqueue = Queue()
-
-    # open threads
-    for i in range(len(channels)):
-        t = ThreadChannelQuery(inqueue, outqueue)
-        t.setDaemon(True)
-        t.start()
-
-    # populate input queue
-    for i, c in enumerate(channels):
-        inqueue.put((i, c))
-
-    # block
-    inqueue.join()
-    outqueue.join()
-    result = []
-    for i in range(len(channels)):
-        c = outqueue.get()
-        if isinstance(c, Exception):
-            raise c
-        else:
-            result.append(c)
-    return zip(*sorted(result, key=lambda (idx, chan): idx))[1]
+    if isinstance(segments, DataQualityFlag):
+        segments = segments.active
+    elif not isinstance(segments, segmentlist):
+        segments = SegmentList([Segment(*x) for x in segments])
+    return f(arg1, segments, *args, **kwargs)
 
 
 def override_sample_rate(channel, rate):
@@ -356,6 +225,7 @@ def find_types(site=None, match=None):
     return conn.find_types(site=site, match=match)
 
 
+@use_segmentlist
 def get_timeseries_dict(channels, segments, config=ConfigParser(),
                         cache=None, query=True, nds='guess', multiprocess=True,
                         frametype=None, statevector=False, return_=True,
@@ -444,6 +314,7 @@ def get_timeseries_dict(channels, segments, config=ConfigParser(),
         return out
 
 
+@use_segmentlist
 def _get_timeseries_dict(channels, segments, config=ConfigParser(),
                          cache=None, query=True, nds='guess', frametype=None,
                          multiprocess=True, return_=True, statevector=False,
@@ -451,8 +322,6 @@ def _get_timeseries_dict(channels, segments, config=ConfigParser(),
     """Internal method to retrieve the data for a set of like-typed
     channels using the :meth:`TimeSeriesDict.read` accessor.
     """
-    if isinstance(segments, DataQualityFlag):
-        segments = segments.active
     channels = map(get_channel, channels)
 
     # set classes
@@ -686,6 +555,7 @@ def _get_timeseries_dict(channels, segments, config=ConfigParser(),
     return out
 
 
+@use_segmentlist
 def get_timeseries(channel, segments, config=ConfigParser(), cache=None,
                    query=True, nds='guess', multiprocess=True,
                    frametype=None, statevector=False, return_=True,
@@ -703,14 +573,13 @@ def get_timeseries(channel, segments, config=ConfigParser(), cache=None,
     return
 
 
+@use_segmentlist
 def get_spectrogram(channel, segments, config=ConfigParser(), cache=None,
                     query=True, nds='guess', format='power', return_=True,
                     multiprocess=True, **fftparams):
     """Retrieve the time-series and generate a spectrogram of the given
     channel
     """
-    if isinstance(segments, DataQualityFlag):
-        segments = segments.active
     channel = get_channel(channel)
 
     # read data for all sub-channels
@@ -750,12 +619,11 @@ def get_spectrogram(channel, segments, config=ConfigParser(), cache=None,
         return out
 
 
+@use_segmentlist
 def _get_spectrogram(channel, segments, config=ConfigParser(), cache=None,
                      query=True, nds='guess', format='power', return_=True,
                      frametype=None, multiprocess=True, method='median-mean',
                      **fftparams):
-    if isinstance(segments, DataQualityFlag):
-        segments = segments.active
     channel = get_channel(channel)
     if format in ['rayleigh']:
         method = format
